ai4ef_train_app/trainapp_python_requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
click==8.1.7
fastapi==0.109.1 # updated for vulnerability reasons
matplotlib==3.8.2
numpy==1.26.2
pandas==2.1.3
protobuf==4.25.1
python-dotenv==1.0.0
torch==2.1.2+cpu # install this to force python to not install the gpu version when installing lightning
pytorch-lightning==2.1.2
scikit-learn==1.3.2
statsmodels==0.14.0
uvicorn==0.24.0.post1
u8darts[torch]
SQLAlchemy==2.0.23
httpx==0.25.2
psycopg2-binary==2.9.9
optuna==3.4.0
pytz==2023.3.post1
seaborn==0.13.0
torchmetrics==1.2.1
psycopg2-binary==2.9.9
dagster==1.6.3 
dagster-cloud==1.6.3 
dagster-webserver==1.6.3
boto3==1.35.80
mlflow==2.19.0